Extract common class prim code
